refactor(dashboard): reduce route and preview modules
